Add license and readme to devtools_server.
