+++ packages/devtools/pubspec.yaml
@@ -12,19 +12,17 @@ environment:
   sdk: '>=2.2.0-dev <3.0.0'
 
 dependencies:
-  args: ^1.5.1
   codemirror: ^0.5.3
   collection: ^1.14.11
+  devtools_server:
+    path: ../devtools_server
   http: ^0.12.0+1
   intl: ^0.15.0
   js: ^0.6.1+1
   meta: ^1.1.0
-  path: ^1.6.0
   pedantic: ^1.4.0
   platform_detect: ^1.3.5
   rxdart: ^0.21.0
-  shelf: ^0.7.4
-  shelf_static: ^0.2.8
   vm_service_lib: ^3.14.3-dev.4
   # We would use local dependencies for these packages if pub publish allowed it.
   octicons_css:
