+++ packages/devtools_server/README.md
@@ -1,0 +1,4 @@
+## What is this?
+
+The is a server that supports [Dart DevTools](https://pub.dartlang.org/packages/devtools).
+This package contains shared code for performing functions with DevTools, like launching it, for example.
