+++ packages/devtools_server/lib/devtools_server.dart
@@ -1,0 +1,5 @@
+// Copyright 2018 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+export 'src/server.dart';
