+++ packages/devtools_server/pubspec.yaml
@@ -1,0 +1,18 @@
+name: devtools_server
+description: A server that supports Dart DevTools.
+
+version: 0.0.1
+
+author: Dart Team <misc@dartlang.org>
+homepage: https://github.com/flutter/devtools
+
+environment:
+  sdk: '>=2.2.0-dev <3.0.0'
+
+dependencies:
+  args: ^1.5.1
+  meta: ^1.1.0
+  path: ^1.6.0
+  pedantic: ^1.4.0
+  shelf: ^0.7.4
+  shelf_static: ^0.2.8
